Add SEXPTYPE inference for byte compiled expression
This commit adds support to infer type of bytecompiled expression.
This is useful to infer the type of expressions inside promises.
Currently, the following types are explicitly inferred -

- NILSXP
- LGLSXP
- LANGSXP
- CLOSXP

Other types are automatically inferred by delegating to existing
code for inferring types.There is scope to do more. R byte compiler
has 124 Opcodes. This implementation supports the most common ones
observed in a sample of vignettes. These also match our expectation
about the kind of values found inside promises. The unsupported types
will appear as BCODESXP. These should be negligible compared to other
objects.

Whenever the inference code fails to infer the type, it prints the
OPCODE. We can get the list of unsupported OPCODE from the logs and
add support from them to get complete coverage.
