+++ src/helpers.cpp
@@ -228,176 +228,6 @@ arglist_t get_arguments(dyntrace_context_t *context, call_id_t  call_id, SEXP op
     return arguments;
 }
 
-string sexp_type_to_string(sexp_type s) {
-    switch (s) {
-        case sexp_type::NIL:
-            return "null";
-        case sexp_type::SYM:
-            return "symbol";
-        case sexp_type::LIST:
-            return "list";
-        case sexp_type::CLOS:
-            return "closure";
-        case sexp_type::ENV:
-            return "environment";
-        case sexp_type::PROM:
-            return "promise";
-        case sexp_type::LANG:
-            return "language";
-        case sexp_type::SPECIAL:
-            return "special";
-        case sexp_type::BUILTIN:
-            return "built-in";
-        case sexp_type::CHAR:
-            return "character";
-        case sexp_type::LGL:
-            return "logical";
-        case sexp_type::INT:
-            return "integer";
-        case sexp_type::REAL:
-            return "real";
-        case sexp_type::CPLX:
-            return "complex";
-        case sexp_type::STR:
-            return "string";
-        case sexp_type::DOT:
-            return "dot";
-        case sexp_type::ANY:
-            return "any";
-        case sexp_type::VEC:
-            return "vector";
-        case sexp_type::EXPR:
-            return "expression";
-        case sexp_type::BCODE:
-            return "byte-code";
-        case sexp_type::EXTPTR:
-            return "external_pointer";
-        case sexp_type::WEAKREF:
-            return "weak_reference";
-        case sexp_type::RAW:
-            return "raw";
-        case sexp_type::S4:
-            return "s4";
-        case sexp_type::OMEGA:
-            return "..."; // This one's made up.
-        default:
-            return "<unknown>";
-    }
-}
-SEXPTYPE sexp_type_to_SEXPTYPE(sexp_type s) {
-    switch (s) {
-        case sexp_type::NIL:
-            return NILSXP;
-        case sexp_type::SYM:
-            return SYMSXP;
-        case sexp_type::LIST:
-            return LISTSXP;
-        case sexp_type::CLOS:
-            return CLOSXP;
-        case sexp_type::ENV:
-            return ENVSXP;
-        case sexp_type::PROM:
-            return PROMSXP;
-        case sexp_type::LANG:
-            return LANGSXP;
-        case sexp_type::SPECIAL:
-            return SPECIALSXP;
-        case sexp_type::BUILTIN:
-            return BUILTINSXP;
-        case sexp_type::CHAR:
-            return CHARSXP;
-        case sexp_type::LGL:
-            return LGLSXP;
-        case sexp_type::INT:
-            return INTSXP;
-        case sexp_type::REAL:
-            return REALSXP;
-        case sexp_type::CPLX:
-            return CPLXSXP;
-        case sexp_type::STR:
-            return STRSXP;
-        case sexp_type::DOT:
-            return DOTSXP;
-        case sexp_type::ANY:
-            return ANYSXP;
-        case sexp_type::VEC:
-            return VECSXP;
-        case sexp_type::EXPR:
-            return EXPRSXP;
-        case sexp_type::BCODE:
-            return BCODESXP;
-        case sexp_type::EXTPTR:
-            return EXTPTRSXP;
-        case sexp_type::WEAKREF:
-            return WEAKREFSXP;
-        case sexp_type::RAW:
-            return RAWSXP;
-        case sexp_type::S4:
-            return S4SXP;
-        case sexp_type::OMEGA:
-            return 69; // This one's made up.
-        default:
-            return -1;
-    }
-}
-
-string sexp_type_to_string_short(sexp_type s) {
-    switch (s) {
-        case sexp_type::NIL:
-            return "NIL";
-        case sexp_type::SYM:
-            return "SYM";
-        case sexp_type::LIST:
-            return "LIST";
-        case sexp_type::CLOS:
-            return "CLOS";
-        case sexp_type::ENV:
-            return "ENV";
-        case sexp_type::PROM:
-            return "PROM";
-        case sexp_type::LANG:
-            return "LANG";
-        case sexp_type::SPECIAL:
-            return "SPECIAL";
-        case sexp_type::BUILTIN:
-            return "BUILTIN";
-        case sexp_type::CHAR:
-            return "CHAR";
-        case sexp_type::LGL:
-            return "LGL";
-        case sexp_type::INT:
-            return "INT";
-        case sexp_type::REAL:
-            return "REAL";
-        case sexp_type::CPLX:
-            return "CPL";
-        case sexp_type::STR:
-            return "STR";
-        case sexp_type::DOT:
-            return "DOT";
-        case sexp_type::ANY:
-            return "ANY";
-        case sexp_type::VEC:
-            return "VEC";
-        case sexp_type::EXPR:
-            return "EXPR";
-        case sexp_type::BCODE:
-            return "BCODE";
-        case sexp_type::EXTPTR:
-            return "EXTPTR";
-        case sexp_type::WEAKREF:
-            return "WEAKREF";
-        case sexp_type::RAW:
-            return "RAW";
-        case sexp_type::S4:
-            return "S4";
-        case sexp_type::OMEGA:
-            return "..."; // This one's made up.
-        default:
-            return "NA";
-    }
-}
-
 string recursive_type_to_string(recursion_type type) {
     switch (type) {
         case recursion_type::MUTUALLY_RECURSIVE:
@@ -412,35 +242,3 @@ string recursive_type_to_string(recursion_type type) {
             return "???????";
     }
 }
-
-string full_sexp_type_to_string(full_sexp_type type) {
-    stringstream result;
-    bool first = true;
-    for (auto iterator = type.begin(); iterator != type.end(); ++iterator) {
-        if (first) {
-            first = false;
-        } else {
-            result << "->";
-        }
-        // result << sexp_type_to_string_short(*iterator);
-        result << sexp_type_to_string(*iterator);
-    }
-    return result.str();
-}
-
-string full_sexp_type_to_number_string(full_sexp_type type) {
-    stringstream result;
-    bool first = true;
-    for (auto iterator = type.begin(); iterator != type.end(); ++iterator) {
-        if (first) {
-            first = false;
-        } else {
-            result << ",";
-        }
-        result << sexp_type_to_SEXPTYPE(*iterator);
-    }
-    return result.str();
-}
-
-
-
